# project/Policlinic/Policlinic.WebApplication/API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Policlinic.Server",
    "version": "1.0"
  },
  "paths": {
    "/api/Appointment": {
      "get": {
        "tags": [
          "Appointment"
        ],
        "summary": "Вернуть все приёмы",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceptionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceptionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceptionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Appointment"
        ],
        "summary": "Добавить новый приём",
        "requestBody": {
          "description": "объект Policlinic.Server.Dto.ReceptionDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceptionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceptionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReceptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Appointment/{id}": {
      "get": {
        "tags": [
          "Appointment"
        ],
        "summary": "Вернуть приём по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор приёма",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReceptionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceptionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceptionDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Appointment"
        ],
        "summary": "Обновить приём по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор приёма",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "объект Policlinic.Server.Dto.ReceptionDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceptionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceptionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReceptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Appointment"
        ],
        "summary": "Удалить приём по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор приёма",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Doctor": {
      "get": {
        "tags": [
          "Doctor"
        ],
        "summary": "Вернуть всех докторов",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Doctor"
        ],
        "summary": "Добавить доктора",
        "requestBody": {
          "description": "объект Policlinic.Server.Dto.DoctorPostDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Doctor/{id}": {
      "get": {
        "tags": [
          "Doctor"
        ],
        "summary": "Вернуть доктора по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор доктора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Doctor"
        ],
        "summary": "Изменить доктора по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор доктора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "объект Policlinic.Server.Dto.DoctorPostDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Doctor"
        ],
        "summary": "Удалить доктора по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор доктора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Patient": {
      "get": {
        "tags": [
          "Patient"
        ],
        "summary": "Вернуть всех пациентов",
        "responses": {
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patient"
        ],
        "summary": "Добавить пациента",
        "requestBody": {
          "description": "объект Policlinic.Server.Dto.PatientPostDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PatientPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Запрос выполнен успешно"
          }
        }
      }
    },
    "/api/Patient/{id}": {
      "get": {
        "tags": [
          "Patient"
        ],
        "summary": "Вернуть пациента по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор пациента",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PatientGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Patient"
        ],
        "summary": "Изменить пациента по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор пациента",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "объект Policlinic.Server.Dto.PatientPostDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PatientPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Запрос выполнен успешно"
          }
        }
      },
      "delete": {
        "tags": [
          "Patient"
        ],
        "summary": "Удалить пациента по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор пациента",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "Успешное удаление или запись не найдена"
          }
        }
      }
    },
    "/api/Requests/experience-doctors": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Вывод всех докторов, опыт которых больше 10 лет",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorGetDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Requests/patients-of-doctor/{id}": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Вывод всех пациентов указанного доктора, сортировка по имени",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Requests/healthy-patients": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Вывод здоровых на данный момент пациентов",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Requests/count-appointment": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Вывод количества приемов пациентов по врачам за месяц",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Requests/disease-top": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Вывод топа заболеваний",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Requests/patients-with-several-appointment": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Вывод пациентов, записанных к нескольким врачам, старше 30 лет, сортировка по дате рождения",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientGetDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DoctorGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "passport": {
            "type": "integer",
            "format": "int64"
          },
          "fio": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "workExperience": {
            "type": "integer",
            "format": "int32"
          },
          "specializationId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DoctorPostDto": {
        "type": "object",
        "properties": {
          "passport": {
            "type": "integer",
            "format": "int64"
          },
          "fio": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "workExperience": {
            "type": "integer",
            "format": "int32"
          },
          "specializationId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PatientGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "passport": {
            "type": "integer",
            "format": "int64"
          },
          "fio": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PatientPostDto": {
        "type": "object",
        "properties": {
          "passport": {
            "type": "integer",
            "format": "int64"
          },
          "fio": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReceptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "dateAndTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "doctorId": {
            "type": "integer",
            "format": "int32"
          },
          "patientId": {
            "type": "integer",
            "format": "int32"
          },
          "conclusion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}